+++ Web/RDD.Web.Tests/Models/AnotherUser.cs
@@ -2,7 +2,7 @@
 
 namespace Rdd.Web.Tests.Models
 {
-    public class AnotherUser : EntityBase<IUser, int>, IUser
+    public class AnotherUser : EntityBase<int>, IUser
     {
         public override int Id { get; set; }
         public override string Name { get; set; }