+++ Web/RDD.Web.Tests/Models/Department.cs
@@ -3,8 +3,9 @@ using System.Collections.Generic;
 
 namespace Rdd.Web.Tests.Models
 {
+    
     public enum Test { A = 0, B = 10 }
-    public class Department : EntityBase<Department, int>
+    public class Department : EntityBase<int>
     {
         public override int Id { get; set; }
         public override string Name { get; set; }