+++ Web/RDD.Web/Controllers/WebController.cs
@@ -15,7 +15,7 @@ using System.Threading.Tasks;
 namespace Rdd.Web.Controllers
 {
     public abstract class WebController<TEntity, TKey> : WebController<IAppController<TEntity, TKey>, TEntity, TKey>
-        where TEntity : class, IEntityBase<TEntity, TKey>
+        where TEntity : class, IEntityBase<TKey>
         where TKey : IEquatable<TKey>
     {
         protected WebController(IAppController<TEntity, TKey> appController, ApiHelper<TEntity, TKey> helper)
@@ -26,7 +26,7 @@ namespace Rdd.Web.Controllers
 
     public abstract class WebController<TAppController, TEntity, TKey> : ReadOnlyWebController<TAppController, TEntity, TKey>
         where TAppController : IAppController<TEntity, TKey>
-        where TEntity : class, IEntityBase<TEntity, TKey>
+        where TEntity : class, IEntityBase<TKey>
         where TKey : IEquatable<TKey>
     {